+++ macros/src/codegen.rs
@@ -111,11 +111,6 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
         })
         .collect();
 
-    // println!("sm: {:#?}", sm);
-    // println!("in_states: {:#?}", in_states);
-    // println!("events: {:#?}", events);
-    // println!("transitions: {:#?}", transitions);
-
     // Map guards, actions and output states into code blocks
     let guards: Vec<Vec<_>> = transitions
         .values()
@@ -252,6 +247,9 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
 
     let mut guard_list = proc_macro2::TokenStream::new();
     let mut action_list = proc_macro2::TokenStream::new();
+
+    let mut entries_exits = proc_macro2::TokenStream::new();
+
     for (state, event_mappings) in transitions.iter() {
         // create the state data token stream
         let state_data = match sm.state_data.data_types.get(state) {
@@ -259,6 +257,18 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
             Some(st) => quote! { state_data: &#st, },
             None => quote! {},
         };
+
+        let entry_ident = format_ident!("on_entry_{}", string_morph::to_snake_case(state));
+        let state_name = format!("[{}::{}]", states_type_name, state);
+        entries_exits.extend(quote! {
+            #[doc = concat!("Called on entry to ", #state_name)]
+            fn #entry_ident(&mut self) {}
+        });
+        let exit_ident = format_ident!("on_exit_{}", string_morph::to_snake_case(state));
+        entries_exits.extend(quote! {
+            #[doc = concat!("Called on exit from ", #state_name)]
+            fn #exit_ident(&mut self) {}
+        });
 
         for (event, event_mapping) in event_mappings {
             for transition in &event_mapping.transitions {
@@ -401,7 +411,7 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
                 .zip(in_states.iter().zip(out_states.iter().zip(guard_action_parameters.iter().zip(guard_action_ref_parameters.iter())))),
         )
         .map(
-            |(guards, (actions, (_, (out_states, (guard_action_parameters, guard_action_ref_parameters)))))| {
+            |(guards, (actions, (in_state, (out_states, (guard_action_parameters, guard_action_ref_parameters)))))| {
                 guards
                     .iter()
                     .zip(
@@ -413,6 +423,20 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
                         let streams: Vec<TokenStream> =
                             guard.iter()
                             .zip(action.iter().zip(out_state)).map(|(guard, (action,out_state))| {
+
+                                let binding = out_state.to_string();
+                                let out_state_string = &binding.split('(').next().unwrap();
+                                let binding = in_state.to_string();
+                                let in_state_string = &binding.split('(').next().unwrap();
+
+                                let entry_ident = format_ident!("on_entry_{}", string_morph::to_snake_case(out_state_string));
+                                let exit_ident = format_ident!("on_exit_{}", string_morph::to_snake_case(in_state_string));
+
+                                let entry_exit_states =
+                                    quote! {
+                                        self.context_mut().#exit_ident();
+                                        self.context_mut().#entry_ident();
+                                        };
                                 let (is_async_action,action_code) = generate_action(action, &temporary_context_call, g_a_param);
                                 is_async_state_machine |= is_async_action;
                                 if let Some(expr) = guard { // Guarded transition
@@ -467,6 +491,7 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
                                               #action_code
                                               let out_state = #states_type_name::#out_state;
                                               self.context.log_state_change(&out_state);
+                                              #entry_exit_states
                                               self.state = Some(out_state);
                                               return self.state()
                                         }
@@ -476,6 +501,7 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
                                        #action_code
                                        let out_state = #states_type_name::#out_state;
                                        self.context.log_state_change(&out_state);
+                                       #entry_exit_states
                                        self.state = Some(out_state);
                                        return self.state();
                                    }
@@ -561,6 +587,8 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
             #guard_error
             #guard_list
             #action_list
+            #entries_exits
+
 
             /// Called at the beginning of a state machine's `process_event()`. No-op by
             /// default but can be overridden in implementations of a state machine's
@@ -679,6 +707,7 @@ pub fn generate_code(sm: &ParsedStateMachine) -> proc_macro2::TokenStream {
                     #states_type_name::#in_states => match event {
                         #(#events_type_name::#events => {
                             #code_blocks
+
                             #[allow(unreachable_code)]
                             {
                                 // none of the guarded or non-guarded transitions occurred,